requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0
streamlit-extras>=0.3.0          # Additional Streamlit components

# Spotify API Integration
spotipy>=2.22.1                  # Spotify Web API Python library
requests>=2.31.0                 # HTTP requests for API calls
requests-oauthlib>=1.3.1         # OAuth handling for Spotify

# Data Processing and Analysis
numpy>=1.24.0                    # Numerical computing
pandas>=2.0.0                    # Data manipulation and analysis
scikit-learn>=1.3.0              # Machine learning (clustering, preprocessing)

# Configuration and Environment Management
python-dotenv>=1.0.0             # Environment variable management
pydantic>=2.4.0                  # Data validation and settings management
pydantic-settings>=2.0.0         # Settings management with Pydantic

# Data Storage and Caching
diskcache>=5.6.3                 # Disk-based caching for API responses

# Audio and Media Processing (Optional)
librosa>=0.10.0                  # Audio analysis (if needed for advanced features)
mutagen>=1.47.0                  # Audio metadata extraction

# Utilities and Helpers
python-dateutil>=2.8.2          # Date/time utilities
tqdm>=4.66.0                     # Progress bars for data processing
tenacity>=8.2.0                  # Retry logic for API calls

# Development and Testing Dependencies
pytest>=7.4.0                   # Testing framework
pytest-cov>=4.1.0               # Test coverage
pytest-mock>=3.11.0             # Mocking for tests
black>=23.0.0                   # Code formatting
flake8>=6.0.0                   # Code linting
mypy>=1.5.0                     # Type checking

# Logging and Monitoring
loguru>=0.7.0                   # Enhanced logging
structlog>=23.1.0               # Structured logging

# Optional: Advanced Features
plotly>=5.17.0                  # Interactive visualizations (for algorithm visualization)
altair>=5.1.0                   # Declarative statistical visualization

# Production Dependencies
gunicorn>=21.2.0                # WSGI server (if deploying beyond Streamlit Cloud)
psutil>=5.9.0                   # System monitoring

# Security
cryptography>=41.0.0            # Cryptographic utilities for secure storage